altered all scripts other than note-deletion.js  and note-search.js to allow note-deletion.js to work properly, Also added deletion confirmation

+++ scripts/for-main-page_dashboard/quick-delete-note.js
@@ -15,39 +15,43 @@ for (i = 0; i < deleteButtons.length; i++) {
 }
 
 function deleteThisNote() {
-
-    //this is used because the button index is shared by its parent element as well as the appropriate note info in the array of stored infos
-    let noteIndex = Array.from(deleteButtons).indexOf(this);
-
-    let notePreviewCardsContainer = document.getElementById("notes-preview-cards-container");
-    let thisNotePreviewCard = this.parentElement;
-
-    //remove from stored note infos
-    storedNoteInfos.splice(noteIndex, 1);
-    console.log("stored note arr", storedNoteInfos);
-    localStorage.setItem("stored-notes", JSON.stringify(storedNoteInfos));
-
-    //start animation and reduce width
-    thisNotePreviewCard.style.flexGrow = "0";
-    thisNotePreviewCard.classList.add("preview-card-exit");
-
-    smoothDeletion();
-
-    function smoothDeletion() {
-        let animationDuration = 300;
-
-        //uses an equation similar to time = distance/speed to determine how long the animation should las and how long to wait before deleting
-        let waitDuration = parseInt(thisNotePreviewCard.clientWidth / animationDuration * 100);
-        thisNotePreviewCard.style.animationDuration = String(waitDuration) + "ms";
-
-        console.log(waitDuration);
-        console.log(thisNotePreviewCard.clientWidth);
-        console.log(animationDuration);
-
-        setTimeout(() => { notePreviewCardsContainer.removeChild(thisNotePreviewCard) }, waitDuration);
+    if (!confirm("This note will be permanently deleted. Would you like to continue?")) {
+        return;
     }
+    else {
+        //this is used because the button index is shared by its parent element as well as the appropriate note info in the array of stored infos
+        let noteIndex = Array.from(deleteButtons).indexOf(this);
 
-    getVariableValues();
+        let notePreviewCardsContainer = document.getElementById("notes-preview-cards-container");
+        let thisNotePreviewCard = this.parentElement;
+
+        //remove from stored note infos
+        storedNoteInfos.splice(noteIndex, 1);
+        console.log("stored note arr", storedNoteInfos);
+        localStorage.setItem("stored-notes", JSON.stringify(storedNoteInfos));
+
+        //start animation and reduce width
+        thisNotePreviewCard.style.flexGrow = "0";
+        thisNotePreviewCard.classList.add("preview-card-exit");
+
+        smoothDeletion();
+
+        function smoothDeletion() {
+            let animationDuration = 300;
+
+            //uses an equation similar to time = distance/speed to determine how long the animation should las and how long to wait before deleting
+            let waitDuration = parseInt(thisNotePreviewCard.clientWidth / animationDuration * 100);
+            thisNotePreviewCard.style.animationDuration = String(waitDuration) + "ms";
+
+            console.log(waitDuration);
+            console.log(thisNotePreviewCard.clientWidth);
+            console.log(animationDuration);
+
+            setTimeout(() => { notePreviewCardsContainer.removeChild(thisNotePreviewCard) }, waitDuration);
+        }
+
+        getVariableValues();
+    }
 }
